refactor(format/yaml): address flow scalar review feedback
