+++ crates/biome_yaml_formatter/src/lib.rs
@@ -1,4 +1,5 @@
 mod comments;
+mod content_lines;
 mod context;
 mod cst;
 mod flow_scalar;
